+++ packages/attention/index.js
@@ -14,6 +14,8 @@ import { messages as daMessages } from './locales/da/messages.mjs';
 import { messages as enMessages } from './locales/en/messages.mjs';
 import { messages as fiMessages } from './locales/fi/messages.mjs';
 import { messages as nbMessages } from './locales/nb/messages.mjs';
+import { messages as svMessages } from './locales/sv/messages.mjs';
+
 import '@warp-ds/icons/elements/close-16';
 
 class WarpAttention extends kebabCaseAttributes(WarpElement) {
@@ -65,7 +67,7 @@ class WarpAttention extends kebabCaseAttributes(WarpElement) {
 
   constructor() {
     super();
-    activateI18n(enMessages, nbMessages, fiMessages, daMessages);
+    activateI18n(enMessages, nbMessages, fiMessages, daMessages, svMessages);
 
     this.handleDone = this.handleDone.bind(this);
 
@@ -98,7 +100,11 @@ class WarpAttention extends kebabCaseAttributes(WarpElement) {
     }
 
     // Fix FOUC effect issues
-    setTimeout(() => this.requestUpdate(), 0);
+    setTimeout(() => {
+      this.requestUpdate();
+      this.handleDone(); // Run handleDone initially, to compute correct arrow position etc. directly.
+    }, 0);
+
     if (!this.callout) {
       window.addEventListener('click', this.handleDone);
       window.addEventListener('scroll', this.handleDone);
@@ -160,7 +166,7 @@ class WarpAttention extends kebabCaseAttributes(WarpElement) {
   }
 
   get _arrowHtml() {
-    return this.noArrow ? '' : html`<div id="arrow" role="img" class="${this._arrowClasses}"></div>`;
+    return this.noArrow ? '' : html`<div id="arrow" class="${this._arrowClasses}"></div>`;
   }
 
   get _activeVariantClasses() {
@@ -362,12 +368,8 @@ class WarpAttention extends kebabCaseAttributes(WarpElement) {
           ? html`
               <slot name="target"></slot>
 
-              <div
-                id="attention"
-                role="${this.tooltip ? 'tooltip' : 'img'}"
-                aria-label="${this.defaultAriaLabel()}"
-                class="${this._wrapperClasses}">
-                ${this._arrowHtml}
+              <div id="attention" class="${this._wrapperClasses}">
+                <div role="${this.tooltip ? 'tooltip' : 'img'}" aria-label="${this.defaultAriaLabel()}">${this._arrowHtml}</div>
                 <slot name="message"></slot>
                 ${this.canClose ? this._closeBtnHtml : nothing}
               </div>
@@ -375,7 +377,8 @@ class WarpAttention extends kebabCaseAttributes(WarpElement) {
           : html`
               <div id="attention" class="${this._wrapperClasses}">
                 <slot name="message"></slot>
-                ${this._arrowHtml} ${this.canClose ? this._closeBtnHtml : nothing}
+                <div role="${this.tooltip ? 'tooltip' : 'img'}" aria-label="${this.defaultAriaLabel()}">${this._arrowHtml}</div>
+                ${this.canClose ? this._closeBtnHtml : nothing}
               </div>
               <slot name="target"></slot>
             `}