+++ packages/toast/locales/da/messages.mjs
@@ -1,1 +1,0 @@
-/*eslint-disable*/export const messages=JSON.parse("{\"toast.aria.error\":\"Fejl\",\"toast.aria.successful\":\"Fuldført\",\"toast.aria.warning\":\"Advarsel\"}");