+++ packages/toast/toast.js
@@ -17,6 +17,7 @@ import { messages as daMessages } from './locales/da/messages.mjs';
 import { messages as enMessages } from './locales/en/messages.mjs';
 import { messages as fiMessages } from './locales/fi/messages.mjs';
 import { messages as nbMessages } from './locales/nb/messages.mjs';
+import { messages as svMessages } from './locales/sv/messages.mjs';
 
 const toastType = {
   success: 'success',
@@ -43,7 +44,7 @@ export class WarpToast extends WarpElement {
 
   constructor() {
     super();
-    activateI18n(enMessages, nbMessages, fiMessages, daMessages);
+    activateI18n(enMessages, nbMessages, fiMessages, daMessages, svMessages);
 
     this.id = Date.now().toString(36) + Math.random().toString(36).slice(2, 5);
     this.type = 'success';